Fix ID types in repositories to match SQL INT type

+++ backend/src/main/kotlin/com/example/testpos/backend/repository/InventoryItemRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface InventoryItemRepository : JpaRepository<InventoryItem, Long>
+interface InventoryItemRepository : JpaRepository<InventoryItem, Int>
